Abanirprint/requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
Deprecated==1.2.18
distro==1.9.0
etelemetry==0.3.1
filelock==3.18.0
frontend==0.0.3
frozenlist==1.7.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
looseversion==1.3.0
lxml==6.0.0
magic-filter==1.0.12
multidict==6.6.3
netaddr==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
propcache==0.3.2
prov==2.1.1
puremagic==1.30
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyxnat==1.6.3
rdflib==7.1.4
reportlab==4.4.3
requests==2.32.4
scipy==1.16.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
WMI==1.5.1
wrapt==1.17.2
yarl==1.20.1
yookassa==3.6.0
